requirements.txt
# Jupyter Lab and core components
jupyterlab==4.2.4
notebook==7.2.1
jupyter-server==2.14.2
ipykernel==6.29.5
jupyter-client==8.6.2

# Core libraries for Spark and data interaction
pyspark==3.4.3
delta-spark==2.4.0
kafka-python-ng  
boto3            

# Source data connection
google-cloud-bigquery
pyarrow          

# Data processing libraries
pandas

# dbt and data transformation
dbt-spark[PyHive]==1.8.0

# Environment variable management
python-dotenv==1.0.1

# Code quality tools
black==24.4.2
flake8==7.0.0